py/dies-py/pyproject.toml
[project]
name = "dies-py"
version = "0.1.0"
dependencies = [
    "msgspec>=0.18.4",
    "pyzmq>=25.1.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = []

[tool.hatch.metadata]
allow-direct-references = true
